Add New assistant button
Added a new assistant button to allow a user to easily add a new assistant.  A file dialog will prompt the user for the name of the file in the default assistants directory.  The newly created assistant will then become the current edited file in the inspector so the configuration could be edited.  The list of available models was converted to an item list so individual models can be selected easily.  A popup will tell the user if they have not selected a model.